+++ src/jsx-runtime.ts
@@ -174,8 +174,8 @@ export interface AntaIntrinsicElements {
   'a-tab-label': BaseAttributes
   'a-tabpanel': ATabpanelAttributes
   'a-steps': BaseAttributes
-  'a-step-ring': BaseAttributes
-  'a-step-copy': BaseAttributes
+  'a-step-marker': BaseAttributes
+  'a-step-desc': BaseAttributes
   'a-step-hint': BaseAttributes
   'a-dialog': ADialogAttributes
   'a-card': ACardAttributes